requirements.txt
#TODO: most likely some dependencies here can be removed
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
Flask==2.3.3
functions-framework==3.4.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-pubsub==2.18.4
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
zipp==3.17.0
